+++ samples/Tutorial/Sequence.fs
@@ -43,8 +43,9 @@ module Cosmos =
     open Equinox.CosmosStore
     let private create (context, cache, accessStrategy) =
         let cacheStrategy = CachingStrategy.SlidingWindow (cache, TimeSpan.FromMinutes 20.) // OR CachingStrategy.NoCaching
-        let category = CosmosStoreCategory(context, Events.codec, Fold.fold, Fold.initial, cacheStrategy, accessStrategy)
-        category |> Equinox.Decider.resolve (Serilog.Log.ForContext<Service>()) |> create
+        CosmosStoreCategory(context, Events.codec, Fold.fold, Fold.initial, cacheStrategy, accessStrategy)
+        |> Equinox.Decider.resolve (Serilog.Log.ForContext<Service>())
+        |> create
 
     module LatestKnownEvent =